inupmt_be/requirements.txt
discord==2.3.2
djangorestframework==3.14.0
drf_yasg
rest-framework-swagger